// src/api/prefect.tsx
export interface components {
    schemas: {
// ...
        Body_count_flow_runs_flow_runs_count_post: {
            flows?: components["schemas"]["FlowFilter"];
            flow_runs?: components["schemas"]["FlowRunFilter"];
            task_runs?: components["schemas"]["TaskRunFilter"];
            deployments?: components["schemas"]["DeploymentFilter"];
            work_pools?: components["schemas"]["WorkPoolFilter"];
            work_pool_queues?: components["schemas"]["WorkQueueFilter"];
        };
// ...
        Body_paginate_flow_runs_flow_runs_paginate_post: {
            /** @default ID_DESC */
            sort: components["schemas"]["FlowRunSort"];
            /**
             * Page
             * @default 1
             */
            page: number;
            flows?: components["schemas"]["FlowFilter"] | null;
            flow_runs?: components["schemas"]["FlowRunFilter"] | null;
            task_runs?: components["schemas"]["TaskRunFilter"] | null;
            deployments?: components["schemas"]["DeploymentFilter"] | null;
            work_pools?: components["schemas"]["WorkPoolFilter"] | null;
            work_pool_queues?: components["schemas"]["WorkQueueFilter"] | null;
            /**
             * Limit
             * @description Defaults to PREFECT_API_DEFAULT_LIMIT if not provided.
             */
            limit?: number;
        };
// ...
        Body_read_flow_run_history_ui_flow_runs_history_post: {
            /** @default EXPECTED_START_TIME_DESC */
            sort: components["schemas"]["FlowRunSort"];
            /**
             * Limit
             * @default 1000
             */
            limit: number;
            /**
             * Offset
             * @default 0
             */
            offset: number;
            flows?: components["schemas"]["FlowFilter"];
            flow_runs?: components["schemas"]["FlowRunFilter"];
            task_runs?: components["schemas"]["TaskRunFilter"];
            deployments?: components["schemas"]["DeploymentFilter"];
            work_pools?: components["schemas"]["WorkPoolFilter"];
        };
// ...
        Body_read_flow_runs_flow_runs_filter_post: {
            /** @default ID_DESC */
            sort: components["schemas"]["FlowRunSort"];
            /**
             * Offset
             * @default 0
             */
            offset: number;
            flows?: components["schemas"]["FlowFilter"] | null;
            flow_runs?: components["schemas"]["FlowRunFilter"] | null;
            task_runs?: components["schemas"]["TaskRunFilter"] | null;
            deployments?: components["schemas"]["DeploymentFilter"] | null;
            work_pools?: components["schemas"]["WorkPoolFilter"] | null;
            work_pool_queues?: components["schemas"]["WorkQueueFilter"] | null;
            /**
             * Limit
             * @description Defaults to PREFECT_API_DEFAULT_LIMIT if not provided.
             */
            limit?: number;
        };
// ...
        Body_set_flow_run_state_flow_runs__id__set_state_post: {
            /** @description The intended state. */
            state: components["schemas"]["StateCreate"];
            /**
             * Force
             * @description If false, orchestration rules will be applied that may alter or prevent the state transition. If True, orchestration rules are not applied.
             * @default false
             */
            force: boolean;
        };
// ...
        DeploymentFlowRunCreate: {
            /** @description The state of the flow run to create */
            state?: components["schemas"]["StateCreate"] | null;
            /**
             * Name
             * @description The name of the flow run. Defaults to a random slug if not specified.
             */
            name?: string;
            /** Parameters */
            parameters?: Record<string, never>;
            /**
             * Enforce Parameter Schema
             * @description Whether or not to enforce the parameter schema on this run.
             */
            enforce_parameter_schema?: boolean | null;
            /** Context */
            context?: Record<string, never>;
            /** Infrastructure Document Id */
            infrastructure_document_id?: string | null;
            /** @description The empirical policy for the flow run. */
            empirical_policy?: components["schemas"]["FlowRunPolicy"];
            /**
             * Tags
             * @description A list of tags for the flow run.
             */
            tags?: string[];
            /**
             * Idempotency Key
             * @description An optional idempotency key. If a flow run with the same idempotency key has already been created, the existing flow run will be returned.
             */
            idempotency_key?: string | null;
            /** Parent Task Run Id */
            parent_task_run_id?: string | null;
            /** Work Queue Name */
            work_queue_name?: string | null;
            /** Job Variables */
            job_variables?: Record<string, never> | null;
        };
// ...
        SimpleFlowRun: {
            /**
             * Id
             * Format: uuid
             * @description The flow run id.
             */
            id: string;
            /** @description The state type. */
            state_type: components["schemas"]["StateType"];
            /**
             * Timestamp
             * Format: date-time
             * @description The start time of the run, or the expected start time if it hasn't run yet.
             */
            timestamp: string;
            /**
             * Duration
             * @description The total run time of the run.
             */
            duration: number;
            /**
             * Lateness
             * @description The delay between the expected and actual start time.
             */
            lateness: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}